Rename `profiling_support.rs` as `self_profile.rs`
It's a clearer name.
