+++ compiler/rustc_query_impl/src/self_profile.rs
